emc.json: remove explicit dependency on Commons Logging 1.2
It turns out that we already have users install Commons Logging 1.2,
as it is directly required by the Excel driver. The problem comes from
how the HttpClient dependency of the EMC JSON driver uses Import-Package
with a very strict [1.1, 1.3) range to obtain Commons Logging.

Xtext from the 2024-12 simrel comes with Commons Logging 1.3.4 which
overrides the Commons Logging 1.2 we install, meaning that HttpClient 4
no longer resolves correctly. We probably just need to upgrade to a
newer version of HttpClient 4 that is in Orbit.