Excel EMC driver: move POI+XMLBeans to core feature, bundle them in feature
We use Maven dependencies to get Apache POI 4.1.2 and the exact version of XMLBeans
used for its generated schemas, and we should ensure the same exact version of POI
and XMLBeans are used to run the driver. The simplest thing is to bundle the wrapped.*
version of XMLBeans with our feature, and change our org.apache.xmlbeans dependency
to the wrapped.org.apache.xmlbeans one with the appropriate version range.