FrameStack: avoid manipulating original collections from the clone
The original version used super.clone(), which will create an object
of the same type with the same field values. This meant that the call
to initializeState() would end up adding another global stack frame
to both the original and the clone. This manifested as more and more
global stack frames appearing every time we evaluated an expression.

This commit changes it to avoid super.clone(), instead creating a
new FrameStack ourselves and using the copy constructors and clone()
methods of the various fields in FrameStack.