ImportManager: importing module should be in cache as well
Suppose we have this situation:

* circular1.egl imports circular2.egl.
* circular2.egl imports circular1.egl.
* We load and parse circular1.egl as the root module.

In this scenario, we want circular1.egl to be in the ImportManager
cache before parsing circular2.egl, so that it will be reused when
parsing circular2.egl.

This prevents the creation of a throwaway EglModule while parsing
circular2.egl (which is not used anyway, due to the loop over the
ancestors which is done in Import::load).